refactor participation into enw module/class;
