+++ CHANGELOG.md
@@ -20,6 +20,7 @@
  - fixes to the dynamic docstrings helper class & setup
  - API simplifications
  - updated flake8 and applied many docstring fixes.
+ - refactored participation methods into module/class.
 
 ## Added
  - new filter package providing classes for filtering solutions
