+++ test/participation/test_fault_system_solution_section_participation.py
@@ -1,6 +1,7 @@
 import pytest
 
 from solvis.filter import FilterRuptureIds, FilterSubsectionIds
+from solvis.solution import SolutionParticipation
 
 RATE_COLUMN = 'rate_weighted_mean'  # 'Annual Rate'
 
@@ -9,7 +10,7 @@ def test_section_participation_rate(crustal_small_fss_fixture):
     # get the participation rate for a single (sub)section
     sec_id = 5
     solution = crustal_small_fss_fixture
-    rates = solution.section_participation_rates([sec_id])
+    rates = SolutionParticipation(solution).section_participation_rates([sec_id])
     print(rates)
     assert pytest.approx(rates.participation_rate.tolist()[0]) == 0.0026206877  # 0.0099396
 
@@ -18,7 +19,7 @@ def test_section_participation_rate_default_spot_check(crustal_small_fss_fixture
     # get the participation rate for a single (sub)section
     sec_id = 5
     solution = crustal_small_fss_fixture
-    rates = solution.section_participation_rates()
+    rates = SolutionParticipation(solution).section_participation_rates()
     # print(f"participation rate for section {sec_id}: {rates['Annual Rate'].tolist()[0]} /yr")
     # rates=r
     assert pytest.approx(rates[rates.index == sec_id].participation_rate.tolist()[0]) == 0.0026206877
@@ -27,7 +28,7 @@ def test_section_participation_rate_default_spot_check(crustal_small_fss_fixture
 def test_section_participation_rate_default_all_sections(crustal_small_fss_fixture):
     # get the participation rate for a single (sub)section
     solution = crustal_small_fss_fixture
-    rates = solution.section_participation_rates()
+    rates = SolutionParticipation(solution).section_participation_rates()
 
     section_rates = solution.model.rs_with_rupture_rates.groupby("section").agg('sum')[RATE_COLUMN]
 
@@ -46,7 +47,7 @@ def test_section_participation_rate_default_all_sections(crustal_small_fss_fixtu
 def test_sum_vs_weighted_mean_all_sections(crustal_small_fss_fixture):
 
     solution = crustal_small_fss_fixture
-    rates = solution.section_participation_rates()
+    rates = SolutionParticipation(solution).section_participation_rates()
     print("rates")
     print(rates)
 
@@ -78,7 +79,7 @@ def test_sum_vs_weighted_mean_conditional(crustal_small_fss_fixture, fault_name,
     rids_subset = rids[2:]
     print(rids_subset)
 
-    srdf = solution.section_participation_rates([subsection_id], rids_subset)
+    srdf = SolutionParticipation(solution).section_participation_rates([subsection_id], rids_subset)
     print("srdf")
     print(srdf)
 
@@ -98,7 +99,7 @@ def test_section_participation_rates(crustal_small_fss_fixture, fault_name, subs
     solution = crustal_small_fss_fixture
 
     subsection_ids = FilterSubsectionIds(solution).for_parent_fault_names([fault_name])
-    srdf = solution.section_participation_rates(subsection_ids)
+    srdf = SolutionParticipation(solution).section_participation_rates(subsection_ids)
     print(srdf)
 
     section_rate = srdf[srdf.index == subsection_id].agg('sum').participation_rate
@@ -106,7 +107,7 @@ def test_section_participation_rates(crustal_small_fss_fixture, fault_name, subs
 
     assert pytest.approx(section_rate) == expected_rate
 
-    srdf2 = solution.section_participation_rates([subsection_id])
+    srdf2 = SolutionParticipation(solution).section_participation_rates([subsection_id])
     print(srdf2)
     assert pytest.approx(srdf2.participation_rate.tolist()[0]) == expected_rate
 
@@ -121,7 +122,7 @@ def test_section_participation_rates_conditional(crustal_small_fss_fixture, faul
     rids_subset = rids[2:]
     print(rids_subset)
 
-    srdf = solution.section_participation_rates([subsection_id], rids_subset)
+    srdf = SolutionParticipation(solution).section_participation_rates([subsection_id], rids_subset)
     print(srdf)
     assert srdf.participation_rate.tolist()[0] - 1e-10 < expected_rate
 
@@ -145,6 +146,6 @@ def test_section_participation_rates_detail(crustal_small_fss_fixture, fault_nam
 
     new_expected_rate = sum(df2[RATE_COLUMN])
 
-    srdf = solution.section_participation_rates([subsection_id], rids_subset)
+    srdf = SolutionParticipation(solution).section_participation_rates([subsection_id], rids_subset)
     print(srdf)
     assert pytest.approx(srdf.participation_rate.tolist()[0]) == new_expected_rate
